Separated unit tests from integration tests
Separated unit tests from integration tests via Maven Failsafe plugin.
Renamed existing integration tests to *IT.java, and left unit tests with
*Test.java.

Running mvn test will execute only unit tests. Running mvn
failsafe:integration-test will run integration tests, which require
proper test_cybs.properties setup.
