+++ src/test/java/com/cybersource/ws/client/ConnectionIT.java
@@ -14,11 +14,14 @@ import org.w3c.dom.Element;
 
 
 /**
- * Test case to validate the HttpClientConnection instance
+ * Test case to validate the Conenction instance
  * @author sunagara
  *
  */
-public class HttpClientConnectionTest {
+public class ConnectionIT {
+
+    private Connection con = null;
+    private Properties merchantProperties = new Properties();
     private String requestFilename = "src/test/resources/auth.xml";
 
     /**
@@ -26,8 +29,6 @@ public class HttpClientConnectionTest {
      */
     @Test
     public void testGetInstance(){
-        HttpClientConnection con = null;
-        Properties merchantProperties = new Properties();
         //Loading the properties file from src/test/resources
         InputStream in = Thread.currentThread().getContextClassLoader().getResourceAsStream("test_cybs.properties");
         if (in == null) {
@@ -38,6 +39,7 @@ public class HttpClientConnectionTest {
         } catch (IOException e1) {
             e1.printStackTrace();
         }
+
         DocumentBuilder builder;
         try {
             builder = Utility.newDocumentBuilder();
@@ -48,16 +50,14 @@ public class HttpClientConnectionTest {
             String nsURI = mc.getEffectiveNamespaceURI();
             setMerchantID(request, merchantID, nsURI);
             LoggerWrapper logger = new LoggerWrapper(null, true, true, mc);
-            con = new HttpClientConnection(mc,builder,logger);
+            con = Connection.getInstance(mc, builder, logger);
             Assert.assertNotNull(con);
-
 
         } catch (ParserConfigurationException e) {
             e.printStackTrace();
         } catch (ConfigException e) {
             e.printStackTrace();
         }
-
     }
 
     /**
